Move some slice methods for better doc order
